ofApp refactor

+++ src/obstacle.h
@@ -6,7 +6,7 @@ class Obstacle {
 	public:
 		Obstacle(int x, int y, int width, int height);
 		void draw();
-		void setWithinWidth(double objectX);
+		void setWithinWidth(float objectX);
 		void lightUp();
 
 		int x, y, width, height, color;
